feat(clients): harmonise evm and chainsauce client builders
When adding another chain, this needed to happen in multiple files. This
is also true for setting a custom fetch client when we need to apss
bearer tokens.

This commit refactors the evmClient and chainsauce builder to rely on a
ChainFactory and RpcFactory to give both clients a consistent config
depending on the selected chain.

Test cases were added and failing tests were updated to respect the new
pattern